app/config.py
import os

from dotenv import load_dotenv

load_dotenv()


DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "postgresql+psycopg2://postgres:postgres@localhost:5432/heb_shared",
)

APP_NAME = "HEB Shared List & Bill Split API"


